refactor: utils
