+++ lib/middleware.js
@@ -10,7 +10,7 @@ const {
   handleRangeHeaders,
   handleRequest,
   ready,
-} = require('./util');
+} = require('./utils');
 
 // Do not add a charset to the Content-Type header of these file types
 // otherwise the client will fail to render them correctly.
