+++ lib/utils/index.js
@@ -1,0 +1,11 @@
+const getFilenameFromUrl = require('./getFilenameFromUrl');
+const handleRangeHeaders = require('./handleRangeHeaders');
+const handleRequest = require('./handleRequest');
+const ready = require('./ready');
+
+module.exports = {
+  getFilenameFromUrl,
+  handleRangeHeaders,
+  handleRequest,
+  ready,
+};
